Implement ReducedPrivileges filesystem support.
This implements a new ReducedPrivileges filesystem type that
changes UID and GID whenever the underlying real filesystem is
accessed.

It also implements new options for the tar() and find() methods
that will limit the number of filesystem changes that are made
during recursion to improve performance with the ReducedPrivileges
filesystem type.

Several minor bugs were also corrected:
 - spurious tar header emitted for unreadable files
 - incorrect handling of 'detached' inodes in Real directories
 - incorrect handling of inodes named '0'
 - alias was copying symlink target rather than the symlink
 - some error states were throwing misformatted errors